Build linearization for `stdlib` only once, when the server starts.
